// src/app/models/CargarSolicitudEstablecimiento.ts
export class CargarSolicitudEstablecimiento {
	catcodigo: string="";
	disdescrip: string="";
	dptdescrip: string="";
	dsacodigo: string="";
	dsadescrip: string="";
	eddcodigo: string="";
	edddescrip: string="DESARROLLO";
	entrazonsoc: string="";
	estnumeins: string="";
	expcodigo: string="";
	expfechreg: string="";
	idexpediente: string="";
	idtramite: string="";
	prvdescrip: string="";
	setcodigo: string="";
	solcodigo: string="";
	solcodiins: string="";
	solcomen: string="";
	soldireccion: string="";
	soldirenume: string="";
	soldirlatitud: string="";
	soldirlongitud: string="";
	solemail: string="";
	solfax: string="";
	solfechreg: string="";
	solfechsol: string="";
	solfechult: string="";
	solindicpc: string="";
	solindipro: string="";
	solindiprodesc: string="";
	solindirpr: string="";
	solinterior: string="";
	sollugareg: string="";
	solnombcom: string="";
	solnumeruc: string="";
	solnumetlf1: string="";
	solnumetlf2: string="";
	solpagweb: string="";
	solrazonsoc: string="";
	solsituasol: string="";
	solurbaniz: string="";
	sstcodigo: string="";
	sstdescrip: string="";
	tprcodigo: string="";
	ttrformatotrm: string="";
	ubicodigo: string="";
	usrdescrip: string="";
	usrlogin: string="";
	LugarRegistro: string = "";
}